Uint8ArrayReader: Possibly performance improvement
The entire file from 0 to the current offset would be traveled due to the `countLineBreaks` function that would be called whenever the `current` method would be called.

We moved the ErrorFormatter.format calls to be called right before the throw, avoiding unnecessary error formatting when no errors are going to be thrown.